requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
python-dotenv==1.0.0
Flask-WTF==1.2.1
email-validator==2.1.1
requests==2.31.0
pytest==8.0.0
